% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_count_matrix}
\alias{gene_count_matrix}
\title{Raw count matrix}
\format{
A data frame with 32833 rows and 16 variables:
#' \describe{
  \item{gene_id}{Gene TAIR ID}
  \item{CFHdF_1}{CFH1 decoy}
  \item{CFHdF_2}{CFH1 decoy}
  \item{CFHdF_3}{CFH1 decoy}
  \item{COL_1}{Columbia wild type}
  \item{COL_2}{Columbia wild type}
  \item{COL_3}{Columbia wild type}
  \item{cfh123_1}{cfh triple mutant}
  \item{cfh123_2}{cfh triple mutant}
  \item{cfh123_3}{cfh triple mutant}
  \item{cfh1_1}{cfh single mutant}
  \item{cfh1_2}{cfh single mutant}
  \item{cfh1_3}{cfh single mutant}
  \item{PIF3OX_1}{PIF3 overexpression}
  \item{PIF3OX_2}{PIF3 overexpression}
  \item{PIF3OX_3}{PIF3 overexpression}
}
}
\usage{
gene_count_matrix
}
\description{
Raw count matrix
}
\keyword{datasets}
