% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_edgeR.R
\name{run_edgeR}
\alias{run_edgeR}
\title{Perform KEGG term enrichment}
\usage{
run_edgeR(contrast, prefix = "")
}
\arguments{
\item{contrast}{Character vector. Column names from the gene_count_matrix.}

\item{prefix}{Character. Add prefix to some of the column names in the output}
}
\value{
Data frame of edgeR result
}
\description{
Perform KEGG term enrichment
}
